requirements.txt
flask==2.0.1
python-dotenv==0.19.0
supabase==1.0.2
stripe==2.70.0
pyjwt==2.3.0
pandas==1.3.4
matplotlib==3.4.3
fpdf2==2.5.5
gunicorn==20.1.0